[world state]: getLabel query now uses integer args, not double

+++ src/worldStateManager/include/WorldStateMgrThread.h
@@ -154,7 +154,7 @@ class WorldStateMgrThread : public RateThread
         bool getTrackerBottleIndexFromID(const int &id, int &tbi);
         bool getAffBottleIndexFromTrackROI(const int &u, const int &v, int &abi);
         int label2id(const string &label);
-        bool getLabel(const double &u, const double &v, string &label);
+        bool getLabel(const int &u, const int &v, string &label);
         bool mono2stereo(const double &u, const double &v, double &x, double &y, double &z);
         // TODO: return Bottle instead of std::vector
         vector<double> getTooltipOffset(const string &objName);
